+++ enterprise/internal/codeintel/autoindexing/internal/background/job_summary_builder.go
@@ -39,7 +39,7 @@ func NewSummaryBuilder(
 				if err != nil {
 					return err
 				}
-				recentIndexes, err := store.GetRecentIndexesSummary(ctx, repositoryWithCount.RepositoryID)
+				recentIndexes, err := uploadSvc.GetRecentIndexesSummary(ctx, repositoryWithCount.RepositoryID)
 				if err != nil {
 					return err
 				}
