+++ enterprise/internal/codeintel/autoindexing/service.go
@@ -20,7 +20,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
-	"github.com/sourcegraph/sourcegraph/internal/symbols"
 	"github.com/sourcegraph/sourcegraph/lib/codeintel/autoindex/config"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
@@ -31,7 +30,6 @@ type Service struct {
 	inferenceSvc    InferenceService
 	repoUpdater     RepoUpdaterClient
 	gitserverClient gitserver.Client
-	symbolsClient   *symbols.Client
 	indexEnqueuer   *enqueuer.IndexEnqueuer
 	jobSelector     *jobselector.JobSelector
 	logger          log.Logger
@@ -45,7 +43,6 @@ func newService(
 	repoUpdater RepoUpdaterClient,
 	repoStore database.RepoStore,
 	gitserverClient gitserver.Client,
-	symbolsClient *symbols.Client,
 ) *Service {
 	// NOTE - this should go up a level in init.go.
 	// Not going to do this now so that we don't blow up all of the
@@ -76,20 +73,11 @@ func newService(
 		repoStore:       repoStore,
 		inferenceSvc:    inferenceSvc,
 		gitserverClient: gitserverClient,
-		symbolsClient:   symbolsClient,
 		indexEnqueuer:   indexEnqueuer,
 		jobSelector:     jobSelector,
 		logger:          observationCtx.Logger,
 		operations:      newOperations(observationCtx),
 	}
-}
-
-func (s *Service) GetRecentIndexesSummary(ctx context.Context, repositoryID int) ([]shared.IndexesWithRepositoryNamespace, error) {
-	return s.store.GetRecentIndexesSummary(ctx, repositoryID)
-}
-
-func (s *Service) GetLastIndexScanForRepository(ctx context.Context, repositoryID int) (*time.Time, error) {
-	return s.store.GetLastIndexScanForRepository(ctx, repositoryID)
 }
 
 func (s *Service) GetIndexConfigurationByRepositoryID(ctx context.Context, repositoryID int) (shared.IndexConfiguration, bool, error) {
@@ -134,7 +122,7 @@ func (s *Service) InferIndexConfiguration(ctx context.Context, repositoryID int,
 		return nil, nil, err
 	}
 
-	indexJobHints, err := s.InferIndexJobHintsFromRepositoryStructure(ctx, repositoryID, commit)
+	indexJobHints, err := s.jobSelector.InferIndexJobHintsFromRepositoryStructure(ctx, repositoryID, commit)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -150,31 +138,6 @@ func (s *Service) InferIndexConfiguration(ctx context.Context, repositoryID int,
 
 func (s *Service) UpdateIndexConfigurationByRepositoryID(ctx context.Context, repositoryID int, data []byte) error {
 	return s.store.UpdateIndexConfigurationByRepositoryID(ctx, repositoryID, data)
-}
-
-func (s *Service) GetSupportedByCtags(ctx context.Context, filepath string, repoName api.RepoName) (bool, string, error) {
-	mappings, err := s.symbolsClient.ListLanguageMappings(ctx, repoName)
-	if err != nil {
-		return false, "", err
-	}
-
-	for language, globs := range mappings {
-		for _, glob := range globs {
-			if glob.Match(filepath) {
-				return true, language, nil
-			}
-		}
-	}
-
-	return false, "", nil
-}
-
-func (s *Service) SetRequestLanguageSupport(ctx context.Context, userID int, language string) error {
-	return s.store.SetRequestLanguageSupport(ctx, userID, language)
-}
-
-func (s *Service) GetLanguagesRequestedBy(ctx context.Context, userID int) ([]string, error) {
-	return s.store.GetLanguagesRequestedBy(ctx, userID)
 }
 
 func (s *Service) QueueRepoRev(ctx context.Context, repositoryID int, rev string) error {
@@ -201,22 +164,18 @@ func (s *Service) InferIndexJobsFromRepositoryStructure(ctx context.Context, rep
 	return s.jobSelector.InferIndexJobsFromRepositoryStructure(ctx, repositoryID, commit, localOverrideScript, bypassLimit)
 }
 
-func (s *Service) InferIndexJobHintsFromRepositoryStructure(ctx context.Context, repositoryID int, commit string) ([]config.IndexJobHint, error) {
-	return s.jobSelector.InferIndexJobHintsFromRepositoryStructure(ctx, repositoryID, commit)
+func IsLimitError(err error) bool {
+	return errors.As(err, &inference.LimitError{})
 }
 
-func (s *Service) NumRepositoriesWithCodeIntelligence(ctx context.Context) (int, error) {
-	return s.store.NumRepositoriesWithCodeIntelligence(ctx)
-}
-
-func (s *Service) RepositoryIDsWithErrors(ctx context.Context, offset, limit int) ([]shared.RepositoryWithCount, int, error) {
-	return s.store.RepositoryIDsWithErrors(ctx, offset, limit)
+func (s *Service) GetRepositoriesForIndexScan(ctx context.Context, table, column string, processDelay time.Duration, allowGlobalPolicies bool, repositoryMatchLimit *int, limit int, now time.Time) ([]int, error) {
+	return s.store.GetRepositoriesForIndexScan(ctx, table, column, processDelay, allowGlobalPolicies, repositoryMatchLimit, limit, now)
 }
 
 func (s *Service) RepositoryIDsWithConfiguration(ctx context.Context, offset, limit int) ([]shared.RepositoryWithAvailableIndexers, int, error) {
 	return s.store.RepositoryIDsWithConfiguration(ctx, offset, limit)
 }
 
-func IsLimitError(err error) bool {
-	return errors.As(err, &inference.LimitError{})
+func (s *Service) GetLastIndexScanForRepository(ctx context.Context, repositoryID int) (*time.Time, error) {
+	return s.store.GetLastIndexScanForRepository(ctx, repositoryID)
 }
