+++ enterprise/internal/codeintel/uploads/internal/store/store_indexes.go
@@ -10,6 +10,7 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 	"go.opentelemetry.io/otel/attribute"
 
+	autoindexingshared "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/autoindexing/shared"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/shared/types"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -332,20 +333,6 @@ func (s *store) GetIndexByID(ctx context.Context, id int) (_ types.Index, _ bool
 	return scanFirstIndex(s.db.Query(ctx, sqlf.Sprintf(getIndexByIDQuery, id, authzConds)))
 }
 
-const indexAssociatedUploadIDQueryFragment = `
-(
-	SELECT MAX(id) FROM lsif_uploads WHERE associated_index_id = u.id
-) AS associated_upload_id
-`
-
-const indexRankQueryFragment = `
-SELECT
-	r.id,
-	ROW_NUMBER() OVER (ORDER BY COALESCE(r.process_after, r.queued_at), r.id) as rank
-FROM lsif_indexes_with_repository_name r
-WHERE r.state = 'queued'
-`
-
 const getIndexByIDQuery = `
 SELECT
 	u.id,
@@ -519,3 +506,138 @@ func scanIndex(s dbutil.Scanner) (index types.Index, err error) {
 
 	return index, nil
 }
+
+// GetRecentIndexesSummary returns the set of "interesting" indexes for the repository with the given identifier.
+// The return value is a list of indexes grouped by root and indexer. In each group, the set of indexes should
+// include the set of unprocessed records as well as the latest finished record. These values allow users to
+// quickly determine if a particular root/indexer pair os up-to-date or having issues processing.
+func (s *store) GetRecentIndexesSummary(ctx context.Context, repositoryID int) (summaries []autoindexingshared.IndexesWithRepositoryNamespace, err error) {
+	ctx, logger, endObservation := s.operations.getRecentIndexesSummary.With(ctx, &err, observation.Args{LogFields: []log.Field{
+		log.Int("repositoryID", repositoryID),
+	}})
+	defer endObservation(1, observation.Args{})
+
+	indexes, err := scanIndexes(s.db.Query(ctx, sqlf.Sprintf(recentIndexesSummaryQuery, repositoryID, repositoryID)))
+	if err != nil {
+		return nil, err
+	}
+	logger.AddEvent("scanIndexes", attribute.Int("numIndexes", len(indexes)))
+
+	groupedIndexes := make([]autoindexingshared.IndexesWithRepositoryNamespace, 1, len(indexes)+1)
+	for _, index := range indexes {
+		if last := groupedIndexes[len(groupedIndexes)-1]; last.Root != index.Root || last.Indexer != index.Indexer {
+			groupedIndexes = append(groupedIndexes, autoindexingshared.IndexesWithRepositoryNamespace{
+				Root:    index.Root,
+				Indexer: index.Indexer,
+			})
+		}
+
+		n := len(groupedIndexes)
+		groupedIndexes[n-1].Indexes = append(groupedIndexes[n-1].Indexes, index)
+	}
+
+	return groupedIndexes[1:], nil
+}
+
+const sanitizedIndexerExpression = `
+(
+    split_part(
+        split_part(
+            CASE
+                -- Strip sourcegraph/ prefix if it exists
+                WHEN strpos(indexer, 'sourcegraph/') = 1 THEN substr(indexer, length('sourcegraph/') + 1)
+                ELSE indexer
+            END,
+        '@', 1), -- strip off @sha256:...
+    ':', 1) -- strip off tag
+)
+`
+
+const indexAssociatedUploadIDQueryFragment = `
+(
+	SELECT MAX(id) FROM lsif_uploads WHERE associated_index_id = u.id
+) AS associated_upload_id
+`
+
+const indexRankQueryFragment = `
+SELECT
+	r.id,
+	ROW_NUMBER() OVER (ORDER BY COALESCE(r.process_after, r.queued_at), r.id) as rank
+FROM lsif_indexes_with_repository_name r
+WHERE r.state = 'queued'
+`
+
+const recentIndexesSummaryQuery = `
+WITH ranked_completed AS (
+	SELECT
+		u.id,
+		u.root,
+		u.indexer,
+		u.finished_at,
+		RANK() OVER (PARTITION BY root, ` + sanitizedIndexerExpression + ` ORDER BY finished_at DESC) AS rank
+	FROM lsif_indexes u
+	WHERE
+		u.repository_id = %s AND
+		u.state NOT IN ('queued', 'processing', 'deleted')
+),
+latest_indexes AS (
+	SELECT u.id, u.root, u.indexer, u.queued_at
+	FROM lsif_indexes u
+	WHERE
+		u.id IN (
+			SELECT rc.id
+			FROM ranked_completed rc
+			WHERE rc.rank = 1
+		)
+	ORDER BY u.root, u.indexer
+),
+new_indexes AS (
+	SELECT u.id
+	FROM lsif_indexes u
+	WHERE
+		u.repository_id = %s AND
+		u.state IN ('queued', 'processing') AND
+		u.queued_at >= (
+			SELECT lu.queued_at
+			FROM latest_indexes lu
+			WHERE
+				lu.root = u.root AND
+				lu.indexer = u.indexer
+			-- condition passes when latest_indexes is empty
+			UNION SELECT u.queued_at LIMIT 1
+		)
+)
+SELECT
+	u.id,
+	u.commit,
+	u.queued_at,
+	u.state,
+	u.failure_message,
+	u.started_at,
+	u.finished_at,
+	u.process_after,
+	u.num_resets,
+	u.num_failures,
+	u.repository_id,
+	u.repository_name,
+	u.docker_steps,
+	u.root,
+	u.indexer,
+	u.indexer_args,
+	u.outfile,
+	u.execution_logs,
+	s.rank,
+	u.local_steps,
+	` + indexAssociatedUploadIDQueryFragment + `,
+	u.should_reindex,
+	u.requested_envvars
+FROM lsif_indexes_with_repository_name u
+LEFT JOIN (` + indexRankQueryFragment + `) s
+ON u.id = s.id
+WHERE u.id IN (
+	SELECT lu.id FROM latest_indexes lu
+	UNION
+	SELECT nu.id FROM new_indexes nu
+)
+ORDER BY u.root, u.indexer
+`
