+++ enterprise/internal/codeintel/uploads/internal/store/store_uploads.go
@@ -424,20 +424,6 @@ func (s *store) GetRecentUploadsSummary(ctx context.Context, repositoryID int) (
 	return groupedUploads[1:], nil
 }
 
-const sanitizedIndexerExpression = `
-(
-    split_part(
-        split_part(
-            CASE
-                -- Strip sourcegraph/ prefix if it exists
-                WHEN strpos(indexer, 'sourcegraph/') = 1 THEN substr(indexer, length('sourcegraph/') + 1)
-                ELSE indexer
-            END,
-        '@', 1), -- strip off @sha256:...
-    ':', 1) -- strip off tag
-)
-`
-
 const recentUploadsSummaryQuery = `
 WITH ranked_completed AS (
 	SELECT
